==> 05b plotting.Rmd <==
---
title: "05 plotting"
output: 
  pdf_document:
    dev: cairo_pdf
date: "2024-12-15"
---

# Description of Code File
This file describes the functions involved in the generation of plots for Analysis 2, which compares the performance of Exogenous vs Non-Exogenous variables, and also Aggregated vs Total admission rates for age and gender.

# Loading the relevant packages
```{r setup, include=FALSE}
rm(list=ls())
library(ggplot2)
library(tidyverse)
library(gridExtra)
library(lubridate)
library(cowplot)
```

# Loading all Data
```{r}
variables = c("age", "gender", "exog")
disease = c("copd","asthma")
for (dis in disease){
  for (var in variables){
    load(paste0(dis, "_", var, "_combined_all_090325.rdata"))
  }
  load(paste0(dis, "_exog_combined_all_090325.rdata"))
  load(paste0(dis, "_noexog_combined_all_090325.rdata"))
}
```

# General Functions
- data_editing, naive_model, and naive_mean_generator are used for computation for MASE

```{r}
data_editing_cases = function(df_filename, outcome){
  combined_df = read_csv(df_filename)%>%
  column_to_rownames(var = "...1")
  
  if (outcome == "cases_copd"){
    popn = "residents_copd"
  } else if (outcome == "cases_asthma"){
    popn = "residents_asthma"
  } else {
    actual_outcome = sub("_([^_]+)$", "", outcome)
    last_part = sub(".*_(.*)$", "\\1", outcome)
    popn = colnames(combined_df)[str_detect(colnames(combined_df), paste0(actual_outcome,"_residents","_",last_part))]
  }
  new_combined = combined_df %>% 
    dplyr::select(year, week, outcome, popn)
  if (outcome == "male_asthma"|outcome=="male_copd"){
    new_combined = new_combined %>% 
      dplyr::select(-starts_with("female_residents"))
  }
  return(new_combined)
}

naive_model = function(df, horizon, outcome){
  df[[outcome]] = dplyr::lag(df[[outcome]],n=horizon)
  df = na.omit(df)
  return(df)
}

naive_mean_generator = function(df_link = "combined_asthma_env_080225_norates.csv", outcome_var, training_limit=0.5){
  df = data_editing_cases(df_link, outcome_var)
  training_length = floor(nrow(df)*training_limit)
  naive_mae = function(df_use, horizon, outcome,training_len){
    naive_x = naive_model(df_use,horizon,outcome = outcome)[[outcome]]
    naive_y = df_use[(horizon+1):nrow(df_use),][[outcome_var]]
    naive_pred = naive_x[4:(training_len+3)]
    naive_actual = naive_y[4:(training_len+3)]
    mean_abs_err = mean(abs(naive_pred-naive_actual))
    return(mean_abs_err)
  }
  
  final_df = data.frame(horizon = numeric(), naive_mean = numeric(), stringsAsFactors = FALSE)
  for (horizon_num in 1:12){
    mae_horizon = naive_mae(df, horizon_num, outcome = outcome_var, training_len = training_length)
    to_add = data.frame(horizon = horizon_num, naive_mean = mae_horizon)
    final_df = rbind(final_df, to_add)
  }
  return(final_df)
}
```

This code chunk shows the MAPE dataframe generator function, which aids in the computation of MAPE across each horizon, with a binary variable indicating if the structural break is included.
```{r}

mape_df_generator = function(df, structural_break){
  if (structural_break == T){
    df_long = df %>%
      dplyr::select(actual, horizon, ends_with("_predictions")) %>%
      pivot_longer(
        cols = -c(actual, horizon), 
        names_to = "model", 
        values_to = "pred"
      ) %>%
      mutate(model = recode(model,
                            "aml_predictions" = "MLP",
                            "ar_predictions" = "AR",
                            "gam_lss_predictions" = "GAMLSS",
                            "gam_xgboost_predictions" = "GAM-XGBoost",
                            "glm_gamsel_predictions" = "GLM-GAMSEL",
                            "bg1_predictions" = "BG1",
                            "bg2_predictions" = "BG2",
                            "eq_weights_predictions" = "SA",
                            "med_weights_predictions" = "MF",
                            "trimmed_mean_predictions" = "TM"
                            )) %>%
      group_by(model, horizon)%>%
      filter(actual != 0) %>%
      summarise(mape = 100*mean(abs(actual - pred) / actual,na.rm=T),.groups="drop")
    return(df_long)
  } else {
    df_long = df %>%
      group_by(horizon) %>%
      slice_head(prop = 0.3) %>%
      dplyr::select(actual, horizon, ends_with("_predictions")) %>%
      pivot_longer(
        cols = -c(actual, horizon), 
        names_to = "model", 
        values_to = "pred"
      ) %>%
      mutate(model = recode(model,
                            "aml_predictions" = "MLP",
                            "ar_predictions" = "AR",
                            "gam_lss_predictions" = "GAMLSS",
                            "gam_xgboost_predictions" = "GAM-XGBoost",
                            "glm_gamsel_predictions" = "GLM-GAMSEL",
                            "bg1_predictions" = "BG1",
                            "bg2_predictions" = "BG2",
                            "eq_weights_predictions" = "SA",
                            "med_weights_predictions" = "MF",
                            "trimmed_mean_predictions" = "TM"
                            )) %>%
      group_by(model, horizon)%>%
      filter(actual != 0) %>%
      summarise(mape = 100*mean(abs(actual - pred) / actual,na.rm=T),.groups="drop")
    return(df_long)
  }
}
```

# Plot 1: Exogenous vs Non-Exogenous Variables
- This code chunk shows the different functions involved in the generation of plots for Exogenous vs Non-Exogenous Variables.

```{r}
model_group_num = c(
  "MLP" = "Unrestricted Models",
  "AR" = "Simple Models",
  "BG1" = "Forecast Combinations",
  "BG2" = "Forecast Combinations",
  "SA" = "Forecast Combinations",
  "GAMLSS" = "Unrestricted Models",
  "GAM-XGBoost" = "Two-Step Forecasting Models",
  "GLM-GAMSEL" = "Two-Step Forecasting Models",
  "MF" = "Forecast Combinations",
  "Naive" = "Simple Models",
  "TM" = "Forecast Combinations"
)

shapes_list = seq(5,19)

mape_plotter = function(df, structural_break = T, title_name, exog_bin = T, dis){
  if (structural_break == T){
    structural_break_bin = "With Structural Break"
  } else {
    structural_break_bin = "Without Structural Break"
  }
  if (exog_bin == T){
    plot_title_size = 85
  } else{
    plot_title_size = 70
  }
  df_long = mape_df_generator(df, structural_break)
  combined_plot_mape = df_long %>% mutate(model_grp = model_group_num[model])
  
  unique_groups = unique(combined_plot_mape$model)
  
  shape_mapping = setNames(shapes_list[seq_along(unique_groups)], unique_groups)
  
  final_plot = ggplot(combined_plot_mape, aes(x = horizon, y = mape, color = model_grp, shape = model)) +
    geom_line(size = 3) +
    geom_point(size = 15, stroke = 3) +
    labs(
      title = paste(title_name),
      x = "Horizon", y = paste0("MAPE (", dis,", %)"), color = "Model Type", shape = "Model") +
    scale_shape_manual(values = shape_mapping) +
    scale_x_continuous(breaks = 1:12, limits = c(1, 12)) +
    scale_y_continuous(breaks = seq(20,100,by = 20), limits = c(15, 100)) +
    theme_minimal() +
    theme(
      plot.title = element_text(hjust = 0.5, size = plot_title_size, face = "bold"),
      legend.position = "right",
      panel.grid = element_blank(),
      axis.title.y = element_text(face = "bold",size=85),
      axis.title.x = element_text(face="bold",size=85),
      panel.border = element_rect(color = "black", fill = NA, size = 1),
      legend.title = element_text(face = "bold", size = 85),
      legend.text = element_text(size=85),
      strip.text = element_text(size = 85),
      axis.text = element_text(size = 80),
      legend.key.size = unit(2, "lines"),
      axis.ticks = element_line(size = 1.5),
      axis.ticks.length = unit(1, "cm")
      )
  return(final_plot)
}

var_name_mapping = c(
  "age_0_to_14" = "Ages 0 to 14",
  "age_15_to_39" = "Ages 15 to 39",
  "age_40_to_54" = "Ages 40 to 54",
  "age_55_to_64" = "Ages 55 to 64",
  "age_65_to_74" = "Ages 65 to 74",
  "age_75_to_84" = "Ages 75 to 84",
  "age_85_above" = "85 and above",
  "discharge" = "Discharge Cases",
  "Female" = "Female",
  "Male" = "Male",
  "age_exog" = "Age (With Exogenous)",
  "age_noexog" = "Age (No Exogenous)",
  "gender_exog" = "Gender (With Exogenous)",
  "gender_noexog" = "Gender (No Exogenous)",
  "exog" = "With Exogenous",
  "noexog" = "No Exogenous"
)

combined_plotter_1 = function(disease, disease_var, row_lab){
  all_plots_sb = list()
  if (disease == "asthma"){
    cap_disease = "Asthma"
  } else{
    cap_disease = "COPD"
  }
  
  for (i in seq_along(disease_var)){
    var = disease_var[i]
    if (i <= 2) { # Specifications to control labelling
      prefix = row_lab
      num = i
    } else if (i > length(disease_var) - 2) {
      prefix = row_lab
      num = i
    }
    df = get(paste0(disease, "_", var, "_combined_all_090325"))
    var_name = paste0(prefix, num, ": ", var_name_mapping[[var]])
    plot = mape_plotter(df, structural_break = T, title_name = var_name, dis = cap_disease) + theme(legend.position = "none")
    
    if (i %% 2 == 0) { # Controls position of y and x axis
      plot = plot + theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())
    }
    if (disease == "copd"){
      plot = plot + theme(axis.title.x = element_blank(), axis.text.x = element_blank(), axis.ticks.x = element_blank())
    }
    plot = plot + theme(plot.margin = margin(t = 50, r = 0, b = 0, l = 0))
    all_plots_sb[[var_name]] = plot
  }
  
  all_plots_nsb = list()
  for (i in seq_along(disease_var)){
    var = disease_var[i]
    
    if (i <= 2) {
      prefix = row_lab
      num = i+2
    } else if (i > 2) {
      prefix = row_lab
      num = i+2
    }
    
    df = get(paste0(disease, "_", var, "_combined_all_090325"))
    var_name = paste0(prefix, num, ": ", var_name_mapping[[var]])
    
    plot = mape_plotter(df, structural_break = F, title_name = var_name, dis = cap_disease) + theme(legend.position = "none") + theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())
    
    if (disease == "copd"){
      plot = plot + theme(axis.title.x = element_blank(), axis.text.x = element_blank(), axis.ticks.x = element_blank())
    }
    plot = plot + theme(plot.margin = margin(t = 50, r = 0, b = 0, l = 0))
    all_plots_nsb[[var_name]] = plot
  }
  
  dashed_line = ggplot() + 
    geom_segment(aes(x = 0, xend = 0, y = -5, yend = 5), 
                 linetype = "dashed", size = 10, color = "black") +
    theme_void() + 
    theme(plot.margin = margin(0, 0, 0, 0, "cm"))
  
  final_plot = plot_grid(plot_grid(plotlist = all_plots_sb,
                        ncol=2,
                        rel_widths = c(1.2,1)),
              dashed_line,
              plot_grid(plotlist = all_plots_nsb,
                        ncol=2,
                        rel_widths = c(1,1)),
              ncol = 3, rel_widths = c(1.05, 0.05, 1))
    
  return(final_plot)
}
```

# Plotting of Exogenous vs Non-Exogenous plots, and saving as Figure 5
```{r}
variables = c("exog", "noexog")

copd_mape_1 = combined_plotter_1("copd", variables,row_lab="A")
legend_1 = get_legend(mape_plotter(asthma_exog_combined_all_090325, structural_break = T, title_name = "", dis = "copd"))
asthma_mape_1 = combined_plotter_1("asthma", variables,row_lab="B")

overall_plot = plot_grid(
  plot_grid(copd_mape_1, asthma_mape_1, nrow=2, rel_heights = c(1,1.1)),
  legend_1,
  ncol = 2, rel_widths = c(4,1)
)
ggsave(file="Figure 5.pdf", plot=overall_plot, width=90, height=40, limitsize=FALSE)
```

# Plot 2: Total Admission Rate vs Aggregated Admission Rates
- This code chunk shows the different functions involved in the generation of plots for Total Admission Rate vs Aggregated Admission Rates.

```{r}

var_name_mapping = c(
  "age_0_to_14" = "Ages 0 to 14",
  "age_15_to_39" = "Ages 15 to 39",
  "age_40_to_54" = "Ages 40 to 54",
  "age_55_to_64" = "Ages 55 to 64",
  "age_65_to_74" = "Ages 65 to 74",
  "age_75_to_84" = "Ages 75 to 84",
  "age_85_above" = "85 and above",
  "discharge" = "Discharge Cases",
  "Female" = "Female",
  "Male" = "Male",
  "exog" = "Total Forecasts",
  "age" = "Age (Aggregated Forecasts)",
  "gender" = "Gender (Aggregated Forecasts)"
)


combined_plotter_2 = function(disease, disease_var, labels){
  all_plots_sb = list()
  if (disease == "asthma"){
    cap_disease = "Asthma"
  } else{
    cap_disease = "COPD"
  }
  
  for (i in seq_along(disease_var)){
    var = disease_var[i]
    df = get(paste0(disease, "_", var, "_combined_all_090325"))
    
    if (disease == "copd"){
      if (var == "exog"){
        prefix = "A"
        num = 1
      } else if (var == "age"){
        prefix = "B"
        num = 1
      } else {
        prefix = "B"
        num = 2
      }
    } else {
      if (var == "exog"){
        prefix = "C"
        num = 1
      } else if (var == "age"){
        prefix = "D"
        num = 1
      } else {
        prefix = "D"
        num = 2
      }
    }
    var_name = paste0(prefix, num, ": ", var_name_mapping[[var]])
    
    # Total Forecast Plot
    plot = mape_plotter(df, structural_break = T, title_name = var_name, exog_bin = F, dis = cap_disease) + theme(legend.position = "none")

    if (i == 3) { # If third element, no need y-axis
      plot = plot + theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())
    }
    if (disease == "copd"){ # COPD plots no need x-axis
      plot = plot + theme(axis.title.x = element_blank())
    } else {
      if (i <= length(disease_var) - 2) {
        plot = plot + theme(axis.title.x = element_blank())
      }
    }
    
    plot = plot + theme(plot.margin = margin(t = 50, r = 0, b = 0, l = 0))
    all_plots_sb[[var_name]] = plot
  }
  
  all_plots_nsb = list()
  for (i in seq_along(disease_var)){
    var = disease_var[i]
    df = get(paste0(disease, "_", var, "_combined_all_090325"))
    
    if (disease == "copd"){
      if (var == "exog"){
        prefix = "A"
        num = 2
      } else if (var == "age"){
        prefix = "B"
        num = 3
      } else {
        prefix = "B"
        num = 4
      }
    } else {
      if (var == "exog"){
        prefix = "C"
        num = 2
      } else if (var == "age"){
        prefix = "D"
        num = 3
      } else {
        prefix = "D"
        num = 4
      }
    }
    var_name = paste0(prefix, num, ": ", var_name_mapping[[var]])
    plot = mape_plotter(df, structural_break = F, title_name = var_name, exog_bin = F, dis = cap_disease) + theme(legend.position = "none") + theme(axis.title.y = element_blank(), axis.text.y = element_blank(), axis.ticks.y = element_blank())
    
    if (disease == "copd"){ # COPD plots no need x-axis titles
      plot = plot + theme(axis.title.x = element_blank())
    } else {
      if (i <= length(disease_var) - 2) {
        plot = plot + theme(axis.title.x = element_blank())
      }
    }
    plot = plot + theme(plot.margin = margin(t = 50, r = 0, b = 0, l = 0))
    all_plots_nsb[[var_name]] = plot
  }
  
  dashed_line = ggplot() + 
    geom_segment(aes(x = 0, xend = 0, y = -5, yend = 5), 
                 linetype = "dashed", size = 10, color = "black") +
    theme_void() + 
    theme(plot.margin = margin(0, 0, 0, 0, "cm"))
  
  final_plot = plot_grid(
    plot_grid(
      all_plots_sb[[1]], 
      plot_grid(all_plots_sb[[2]], all_plots_sb[[3]], ncol = 2, rel_widths = c(1.2,1)), # Second row
      nrow = 2,
      rel_heights = c(1, 1.1)),
    dashed_line,
    plot_grid(
      all_plots_nsb[[1]], 
      plot_grid(all_plots_nsb[[2]], all_plots_nsb[[3]], ncol = 2), # Second row
      nrow = 2,
      rel_heights = c(1, 1.1)),
    ncol = 3, rel_widths = c(1.05, 0.05, 1))
  return(final_plot)
}
```

# Plotting of Total Admission Rate vs Aggregated Admission Rate plots, and saving as Figure 6
```{r}
variables = c("exog", "age", "gender")

copd_mape = combined_plotter_2("copd", variables, labels = c("A","B"))
asthma_mape = combined_plotter_2("asthma", variables, labels = c("C","D"))
legend = get_legend(mape_plotter(asthma_exog_combined_all_090325, structural_break = T, title_name = "", dis = "asthma"))
overall_plot = plot_grid(
  plot_grid(copd_mape, asthma_mape, nrow=2, rel_heights = c(1,1.1)),
  legend,
  ncol = 2, rel_widths = c(4,1)
)
ggsave(file="Figure 6.pdf", plot=overall_plot, width=90, height=75, limitsize=FALSE)

```

# Dataframe generator functions for MAE and MASE
- This generator functions aid in summarising the forecasts to compute MAE and MASE necessary for the SI

```{r}

mae_df_generator = function(df, with_structural_break){
  if (with_structural_break == T){
    df_long = df %>%
      dplyr::select(actual, horizon, ends_with("_predictions")) %>%
      pivot_longer(
        cols = -c(actual, horizon), 
        names_to = "model", 
        values_to = "pred"
      ) %>%
      mutate(pred = ifelse(pred<0,0,pred)) %>%
      mutate(model = recode(model,
                            "aml_predictions" = "MLP",
                            "ar_predictions" = "AR",
                            "gam_lss_predictions" = "GAMLSS",
                            "gam_xgboost_predictions" = "GAM-XGBoost",
                            "glm_gamsel_predictions" = "GLM-GAMSEL",
                            "bg1_predictions" = "BG1",
                            "bg2_predictions" = "BG2",
                            "eq_weights_predictions" = "SA",
                            "med_weights_predictions" = "MF",
                            "trimmed_mean_predictions" = "TM"
                            )) %>%
      group_by(model, horizon) %>%
      summarise(mae = mean(abs(actual-pred)),.groups="drop")
    return(df_long)
    
  } else {
    df_long= df %>%
      group_by(horizon) %>%
      slice_head(prop = 0.3) %>%
      ungroup() %>%
      dplyr::select(actual, horizon, ends_with("_predictions")) %>%
      pivot_longer(
        cols = -c(actual, horizon), 
        names_to = "model", 
        values_to = "pred"
      ) %>%
      mutate(pred = ifelse(pred<0,0,pred)) %>%
      mutate(model = recode(model,
                            "aml_predictions" = "MLP",
                            "ar_predictions" = "AR",
                            "gam_lss_predictions" = "GAMLSS",
                            "gam_xgboost_predictions" = "GAM-XGBoost",
                            "glm_gamsel_predictions" = "GLM-GAMSEL",
                            "bg1_predictions" = "BG1",
                            "bg2_predictions" = "BG2",
                            "eq_weights_predictions" = "SA",
                            "med_weights_predictions" = "MF",
                            "trimmed_mean_predictions" = "TM"
                            )) %>%
      group_by(model, horizon) %>%
      summarise(mae = mean(abs(actual-pred)),.groups="drop")
    return(df_long)
  }
}

mase_df_generator = function(df, naive_mean_df, structural_break){
  if (structural_break == T){
    df_long = df %>%
      dplyr::select(actual, horizon, ends_with("_predictions")) %>%
      pivot_longer(
        cols = -c(actual, horizon), 
        names_to = "model", 
        values_to = "pred"
      ) %>%
      mutate(model = recode(model,
                            "aml_predictions" = "MLP",
                            "ar_predictions" = "AR",
                            "gam_lss_predictions" = "GAMLSS",
                            "gam_xgboost_predictions" = "GAM-XGBoost",
                            "glm_gamsel_predictions" = "GLM-GAMSEL",
                            "bg1_predictions" = "BG1",
                            "bg2_predictions" = "BG2",
                            "eq_weights_predictions" = "SA",
                            "med_weights_predictions" = "MF",
                            "trimmed_mean_predictions" = "TM"
                            )) %>%
      group_by(model, horizon) %>%
      left_join(naive_mean_df, by="horizon") %>%
      summarise(mase = mean(abs(actual-pred))/naive_mean,.groups="drop") %>%
      unique()
    return(df_long)
  } else {
    df_long= df %>%
      group_by(horizon) %>%
      slice_head(prop = 0.3) %>%
      ungroup() %>%
      dplyr::select(actual, horizon, ends_with("_predictions")) %>%
      pivot_longer(
        cols = -c(actual, horizon), 
        names_to = "model", 
        values_to = "pred"
      ) %>%
      mutate(model = recode(model,
                            "aml_predictions" = "MLP",
                            "ar_predictions" = "AR",
                            "gam_lss_predictions" = "GAMLSS",
                            "gam_xgboost_predictions" = "GAM-XGBoost",
                            "glm_gamsel_predictions" = "GLM-GAMSEL",
                            "bg1_predictions" = "BG1",
                            "bg2_predictions" = "BG2",
                            "eq_weights_predictions" = "SA",
                            "med_weights_predictions" = "MF",
                            "trimmed_mean_predictions" = "TM"
                            )) %>%
      group_by(model, horizon) %>%
      left_join(naive_mean_df, by="horizon") %>%
      summarise(mase = mean(abs(actual-pred))/naive_mean,.groups="drop") %>%
      unique()
    return(df_long)
  }
}
```

# Data: MAE Data Tables
Supplementary Information S1 Data Tables

```{r}
csv_generator = function(df, with_structural_break){
  err_metrics = mae_df_generator(df, with_structural_break) %>% 
      pivot_wider(
        names_from = model,   
        values_from = mae) %>% 
      mutate(across(where(is.numeric), ~ round(., 2)))
  if (with_structural_break == T){
    
    mean_values = df %>% 
      group_by(horizon) %>% 
      summarise(mean_case = signif(mean(actual),3),.groups="drop")
    final_df = err_metrics %>%
      inner_join(mean_values, by = "horizon")
  } else {
    mean_values = df %>% 
      group_by(horizon) %>% 
      slice_head(prop = 0.3) %>%
      summarise(mean_case = signif(mean(actual),3),.groups="drop")
    final_df = err_metrics %>%
      inner_join(mean_values, by = "horizon")
  }
  final_df = final_df %>% 
    dplyr::select(horizon, AR, GAMLSS, `GAM-XGBoost`, `GLM-GAMSEL`, MLP, BG1, BG2, SA, MF, TM, mean_case)
  return(final_df)
}

# Analysis 1

variables_1 = c("exog", "noexog")

for (var in variables_1){
  file_name = paste0("asthma_",var,"_combined_ensem_240125")
  file = get(file_name)
  df_struc = csv_generator(file, with_structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("asthma_mae_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  # print(save_path)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = csv_generator(file, with_structural_break = F)
  output_file_name = paste0("asthma_mae_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}

for (var in variables_1){
  file_name = paste0("copd_",var,"_combined_ensem_240125")
  file = get(file_name)
  df_struc = csv_generator(file, with_structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("copd_mae_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = csv_generator(file, with_structural_break = F)
  output_file_name = paste0("copd_mae_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}

# Analysis 2
final_var = c("age_exog", "age", "gender_exog", "gender")

for (var in final_var){
  file_name = paste0("asthma_",var,"_combined_200125")
  file = get(file_name)
  df_struc = csv_generator(file, with_structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("asthma_mae_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = csv_generator(file, with_structural_break = F)
  output_file_name = paste0("asthma_mae_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}

for (var in final_var){
  file_name = paste0("copd_",var,"_combined_200125")
  file = get(file_name)
  df_struc = csv_generator(file, with_structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("copd_mae_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = csv_generator(file, with_structural_break = F)
  output_file_name = paste0("copd_mae_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}
```

# Data: MAPE Data Tables
Supplementary Information MAPE S1 Data Tables

```{r}
# Generation of Data Tables
mape_csv = function(df, structural_break){
  final = mape_df_generator(df, structural_break) %>% 
      mutate(mape = mape/100) %>%
      pivot_wider(
        names_from = model,   
        values_from = mape) %>% 
      mutate(across(where(is.numeric), ~ round(., 2))) %>% 
      dplyr::select(horizon, AR, GAMLSS, `GAM-XGBoost`, `GLM-GAMSEL`, MLP, BG1, BG2, SA, MF, TM)
  return(final)
}

# mape_csv(copd_age_exog_combined_200125, structural_break = T)

# Analysis 1
variables_1 = c("exog", "noexog")

for (var in variables_1){
  file_name = paste0("asthma_",var,"_combined_ensem_240125")
  file = get(file_name)
  df_struc = mape_csv(file, structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("asthma_mape_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = mape_csv(file, structural_break = F)
  output_file_name = paste0("asthma_mape_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}

for (var in variables_1){
  file_name = paste0("copd_",var,"_combined_ensem_240125")
  file = get(file_name)
  df_struc = mape_csv(file, structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("copd_mape_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = mape_csv(file, structural_break = F)
  output_file_name = paste0("copd_mape_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}

# Analysis 2
final_var = c("age_exog", "age", "gender_exog", "gender", "age_noexog", "gender_noexog")

for (var in final_var){
  file_name = paste0("asthma_",var,"_combined_200125")
  file = get(file_name)
  df_struc = mape_csv(file, structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("asthma_mape_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = mape_csv(file, structural_break = F)
  output_file_name = paste0("asthma_mape_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}

for (var in final_var){
  file_name = paste0("copd_",var,"_combined_200125")
  file = get(file_name)
  df_struc = mape_csv(file, structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("copd_mape_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = mape_csv(file, structural_break = F)
  output_file_name = paste0("copd_mape_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}
```

# Data: MASE Data Tables
Supplementary Information S1 Data Tables

```{r}
mase_csv = function(df, naive_mean_df, structural_break){
  final = mase_df_generator(df, naive_mean_df, structural_break) %>% 
      pivot_wider(
        names_from = model,   
        values_from = mase) %>% 
      mutate(across(where(is.numeric), ~ round(., 2))) %>% 
      dplyr::select(horizon, AR, GAMLSS, `GAM-XGBoost`, `GLM-GAMSEL`, MLP, BG1, BG2, SA, MF, TM)
  return(final)
}

# Analysis 1

variables_1 = c("exog", "noexog")

for (var in variables_1){
  file_name = paste0("asthma_",var,"_combined_ensem_240125")
  
  naive_mean_df = naive_mean_generator(outcome_var = "overall_rate_asthma")
  file = get(file_name)
  df_struc = mase_csv(file, naive_mean_df, structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("asthma_mase_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = mase_csv(file, naive_mean_df, structural_break = F)
  output_file_name = paste0("asthma_mase_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}

for (var in variables_1){
  file_name = paste0("copd_",var,"_combined_ensem_240125")
  naive_mean_df = naive_mean_generator(outcome_var = "overall_rate_copd")
  file = get(file_name)
  df_struc = mase_csv(file, naive_mean_df, structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("copd_mase_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = mase_csv(file, naive_mean_df, structural_break = F)
  output_file_name = paste0("copd_mase_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}

# Analysis 2

final_var = c("age_exog", "age", "gender_exog", "gender", "age_noexog", "gender_noexog")

for (var in final_var){
  file_name = paste0("copd_",var,"_combined_200125")
  if (str_detect(var,"age")){
    outcome = paste0("discharge_rate_copd", "_age")
  } else {
    outcome = paste0("discharge_rate_copd", "_gender")
  }
  naive_mean_df = naive_mean_generator(outcome_var = outcome)
  file = get(file_name)
  df_struc = mase_csv(file, naive_mean_df, structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("copd_mase_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = mase_csv(file, naive_mean_df, structural_break = F)
  output_file_name = paste0("copd_mase_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}

for (var in final_var){
  file_name = paste0("asthma_",var,"_combined_200125")
  if (str_detect(var,"age")){
    outcome = paste0("discharge_rate_copd", "_age")
  } else {
    outcome = paste0("discharge_rate_copd", "_gender")
  }
  naive_mean_df = naive_mean_generator(outcome_var = outcome)
  file = get(file_name)
  df_struc = mase_csv(file, naive_mean_df, structural_break = T)
  folder_name = paste0(var,"_err")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  output_file_name = paste0("asthma_mase_", var, "_processed_with_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_struc, file = full_save_path, row.names = FALSE)
  
  df_no_struc = mase_csv(file, naive_mean_df, structural_break = F)
  output_file_name = paste0("asthma_mase_", var, "_processed_no_structuralbreak.csv")
  full_save_path = file.path(save_path, output_file_name)
  write.csv(df_no_struc, file = full_save_path, row.names = FALSE)
}
```
