==> 06c S3 generation 110325.Rmd <==
---
title: "Plotter"
output: html_document
date: "2025-01-10"
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
rm(list=ls())
library(tidyverse)
library(stringr)
```

# Analysis 1

```{r}
asthma_variables = c("cases","female", "male", "age_15_to_39", "age_40_to_54", "age_55_to_64", "age_65_to_74", "age_75_to_84", "age_85_above")
copd_variables = c("cases", "female", "male", "age_55_to_64", "age_65_to_74", "age_75_to_84", "age_85_above")
load("copd_cases_all_280225.rdata")

recoder_fn = function(df){
  df_pred = df %>% 
    dplyr::select(index, horizon, actual, ends_with("predictions")) %>%
    rename(
      "MLP" = "aml_predictions",
      "AR (C)" = "ar_env_predictions",
      "AR (D)" = "ar_mixed_predictions",
      "AR (B)" = "ar_other_disease_predictions",
      "BG1" = "bg1_predictions",
      "BG2" = "bg2_predictions",
      "SA" = "eq_weights_predictions",
      "GAMLSS" = "gam_lss_predictions",
      "GAM-XGBoost" = "gam_xgboost_predictions",
      "GLM-GAMSEL" = "glm_gamsel_predictions",
      "MF" = "med_weights_predictions",
      "Naive" = "naive_predictions",
      "AR (A)" = "pure_ar_predictions",
      "TM" = "trimmed_mean_predictions"
    ) %>%
    dplyr::select(-c(index,horizon,actual))
  return(colnames(df_pred))
}

model_names = recoder_fn(copd_cases_all_280225)

plotting_fn = function(df, type_of_response, horizon_num, model_name){
  if (type_of_response == "cases"){
    resp = "Admission Case Counts"
  } else {
    resp = "Admission Rate"
  }
  df_pred = df %>% 
    dplyr::select(index, horizon, actual, ends_with("predictions")) %>%
    rename(
      "MLP" = "aml_predictions",
      "AR (C)" = "ar_env_predictions",
      "AR (D)" = "ar_mixed_predictions",
      "AR (B)" = "ar_other_disease_predictions",
      "BG1" = "bg1_predictions",
      "BG2" = "bg2_predictions",
      "SA" = "eq_weights_predictions",
      "GAMLSS" = "gam_lss_predictions",
      "GAM-XGBoost" = "gam_xgboost_predictions",
      "GLM-GAMSEL" = "glm_gamsel_predictions",
      "MF" = "med_weights_predictions",
      "Naive" = "naive_predictions",
      "AR (A)" = "pure_ar_predictions",
      "TM" = "trimmed_mean_predictions"
    )
  new_name = paste0(model_name, "'s forecasts for ", horizon_num,"-week ahead horizon")
  # return(new_name)
  df_filtered = df_pred %>% 
    filter(horizon == horizon_num) %>% 
    dplyr::select(index, actual, model_name) %>%
    setNames(c("index", "Observed", new_name))
  # return(df_filtered)
  
  plot = ggplot(df_filtered, aes(x=index)) +
    geom_line(aes(y=Observed, color = "Observed")) +
    geom_point(aes(y=!!sym(new_name), color = new_name)) +
    scale_y_continuous(limits = c(0,floor(max(df_filtered$Observed)+0.3*max(df_filtered$Observed)))) +
    scale_color_manual(values = c("black","red"), breaks=c("Observed", new_name)) +
    labs(x = "Test Set Week",
         y = resp) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          legend.key.size = unit(2, "lines"),
          plot.margin = unit(c(1, 1, 1, 1), "cm"),
          axis.title.y = element_text(face = "bold",size=14),
          axis.title.x = element_text(face="bold",size=14),
          panel.border = element_rect(color = "black", fill = NA, size = 1),
          # legend.title = element_text(face = "bold",size=14),
          legend.text = element_text(size=14),
          legend.position = c(0, 1),  
          legend.justification = c(0, 1),
          strip.text = element_text(size = 14),
          axis.text = element_text(size = 14)) +
    guides(color = guide_legend(title = NULL, order=1))
  return(plot)
}

for (var in copd_variables){
  file_name = paste0("copd_",var,"_all_280225.rdata")
  load(file_name)
  df = get(paste0("copd_",var,"_all_280225"))
  folder_name = paste0("copd_",var,"_all")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  if (str_detect("cases",var)){
    var_type = "cases"
  } else {
    var_type = "others"
  }
  for (horizon_n in 1:12){
    for (model in model_names){
      c = plotting_fn(df, type_of_response = var_type, horizon_num = horizon_n, model)
      file_name = paste0(model, "_horizon", horizon_n, ".pdf")
      full_path = file.path(save_path, file_name)
      ggsave(c, filename = full_path, width = 10, height = 4.5)
    }
  }
}

for (var in asthma_variables){
  file_name = paste0("asthma_",var,"_all_280225.rdata")
  load(file_name)
  df = get(paste0("asthma_",var,"_all_280225"))
  folder_name = paste0("asthma_",var,"_all")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  if (str_detect("cases",var)){
    var_type = "cases"
  } else {
    var_type = "others"
  }
  for (horizon_n in 1:12){
    for (model in model_names){
      c = plotting_fn(df, type_of_response = var_type, horizon_num = horizon_n, model)
      file_name = paste0(model, "_horizon", horizon_n, ".pdf")
      full_path = file.path(save_path, file_name)
      ggsave(c, filename = full_path, width = 10, height = 4.5)
    }
  }
}
```

# Analysis 2a

```{r}

final_var_1 = c("exog", "noexog", "gender", "age")
load("copd_exog_combined_all_090325.rdata")

copd_exog_combined_all_090325 = copd_exog_combined_all_090325 
recoder_fn = function(df){
  df_pred = df %>% 
    dplyr::select(horizon, actual, ends_with("predictions")) %>%
    rename(
      "MLP" = "aml_predictions",
      "AR" = "ar_predictions",
      "BG1" = "bg1_predictions",
      "BG2" = "bg2_predictions",
      "SA" = "eq_weights_predictions",
      "GAMLSS" = "gam_lss_predictions",
      "GAM-XGBoost" = "gam_xgboost_predictions",
      "GLM-GAMSEL" = "glm_gamsel_predictions",
      "MF" = "med_weights_predictions",
      "TM" = "trimmed_mean_predictions"
    ) %>%
    dplyr::select(-c(horizon,actual))
  return(colnames(df_pred))
}

model_names = recoder_fn(asthma_exog_combined_all_090325)

plotting_fn_1 = function(df, type_of_response, exog, agg = F, horizon_num, model_name, dis){
  if (type_of_response == "cases"){
    resp = "Admission Case Counts"
  } else {
    resp = paste0(dis, " Admission Rate ")
  }
  if (agg == T){
    agg_str = " Agg "
    exog_str = ""
  } else{
    agg_str = " Overall "
  }
  if (exog == T){
    exog_str = " (with Exog Variables) "
  }
  
  df_pred = df %>% 
    dplyr::select(index, horizon, actual, ends_with("predictions")) %>%
    rename(
      "MLP" = "aml_predictions",
      "AR" = "ar_predictions",
      "BG1" = "bg1_predictions",
      "BG2" = "bg2_predictions",
      "SA" = "eq_weights_predictions",
      "GAMLSS" = "gam_lss_predictions",
      "GAM-XGBoost" = "gam_xgboost_predictions",
      "GLM-GAMSEL" = "glm_gamsel_predictions",
      "MF" = "med_weights_predictions",
      "TM" = "trimmed_mean_predictions"
    )
  new_name = paste0(model_name, "'s", agg_str, resp, "forecasts ", horizon_num, "-week(s) ahead", exog_str)
  
  df_filtered = df_pred %>% 
    filter(horizon == horizon_num) %>% 
    dplyr::select(index, actual, model_name) %>%
    setNames(c("index", "Observed", new_name))
  
  plot = ggplot(df_filtered, aes(x=index)) +
    geom_line(aes(y=Observed, color = "Observed")) +
    geom_point(aes(y=!!sym(new_name), color = new_name)) +
    scale_y_continuous(limits = c(0,floor(max(df_filtered$Observed)+0.3*max(df_filtered$Observed)))) +
    scale_color_manual(values = c("black","red"), breaks=c("Observed", new_name)) +
    labs(x = "Test Set Week",
         y = resp) +
    theme_minimal() +
    theme(panel.grid = element_blank(),
          legend.key.size = unit(2, "lines"),
          plot.margin = unit(c(1, 1, 1, 1), "cm"),
          axis.title.y = element_text(face = "bold",size=14),
          axis.title.x = element_text(face="bold",size=14),
          panel.border = element_rect(color = "black", fill = NA, size = 1),
          legend.text = element_text(size=14),
          legend.position = c(0, 1),  
          legend.justification = c(0, 1),
          strip.text = element_text(size = 14),
          axis.text = element_text(size = 14)) +
    guides(color = guide_legend(title = NULL, order=1))
  return(plot)
}

for (var in final_var_1){
  if (var == "age" | var == "gender"){
    agg_bin = T
    exog_bin = T
  } else if (str_detect(var, "exog")){
    exog_bin = T
    agg_bin = F
  } else {
    exog_bin = F
    agg_bin = F
  }
  
  file_name = paste0("copd_",var,"_combined_all_090325.rdata")
  load(file_name)
  df = get(paste0("copd_",var,"_combined_all_090325"))
  df_new = df %>% group_by(horizon) %>% mutate(index = row_number()) %>% ungroup()
  folder_name = paste0("copd_",var,"_all")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  if (str_detect("cases",var)){
    var_type = "cases"
  } else {
    var_type = "others"
  }
  for (horizon_n in 1:12){
    for (model in model_names){
      c = plotting_fn_1(df_new, type_of_response = var_type, exog = exog_bin, agg = agg_bin, horizon_num = horizon_n, model, dis = "COPD")
      file_name = paste0(model, "_horizon", horizon_n, ".pdf")
      full_path = file.path(save_path, file_name)
      ggsave(c, filename = full_path, width = 10, height = 4.5)
    }
  }
}

for (var in final_var_1){
  if (var == "age" | var == "gender"){
    agg_bin = T
    exog_bin = T
  } else if (str_detect(var, "exog")){
    exog_bin = T
    agg_bin = F
  } else {
    exog_bin = F
    agg_bin = F
  }

  file_name = paste0("asthma_",var,"_combined_all_090325.rdata")
  load(file_name)
  df = get(paste0("asthma_",var,"_combined_all_090325"))
  df_new = df %>% group_by(horizon) %>% mutate(index = row_number()) %>% ungroup()
  folder_name = paste0("asthma_",var,"_all")
  if (!dir.exists(folder_name)) {
    dir.create(folder_name)
  }
  save_path = file.path(getwd(), folder_name)
  if (str_detect("admission",var)){
    var_type = "cases"
  } else {
    var_type = "others"
  }
  for (horizon_n in 1:12){
    for (model in model_names){
      c = plotting_fn_1(df_new, type_of_response = var_type, exog = exog_bin, agg = agg_bin, horizon_num = horizon_n, model, dis = "Asthma")
      file_name = paste0(model, "_horizon", horizon_n, ".pdf")
      full_path = file.path(save_path, file_name)
      ggsave(c, filename = full_path, width = 10, height = 4.5)
    }
  }
}
```
